Move texture tracking into a module
